inbound: split detect module
